Issues#11の対応
