+++ portal/src/main/java/com/portal/z/common/controller/GlobalControllAdvice.java
@@ -1,0 +1,43 @@
+package com.portal.z.common.controller;
+
+import org.springframework.dao.DataAccessException;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Component;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+@Component
+public class GlobalControllAdvice {
+
+    @ExceptionHandler(DataAccessException.class)
+    public String dataAccessExceptionHandler(DataAccessException e, Model model) {
+
+        // 例外クラスのメッセージをModelに登録
+        model.addAttribute("error", "内部サーバーエラー（DB）：GlobalControllAdvice");
+
+        // 例外クラスのメッセージをModelに登録
+        model.addAttribute("message", "DataAccessExceptionが発生しました");
+
+        // HTTPのエラーコード（500）をModelに登録
+        model.addAttribute("status", HttpStatus.INTERNAL_SERVER_ERROR);
+
+        return "error";
+    }
+
+    @ExceptionHandler(Exception.class)
+    public String exceptionHandler(Exception e, Model model) {
+
+        // 例外クラスのメッセージをModelに登録
+        model.addAttribute("error", "内部サーバーエラー：GlobalControllAdvice");
+
+        // 例外クラスのメッセージをModelに登録
+        model.addAttribute("message", "Exceptionが発生しました");
+
+        // HTTPのエラーコード（500）をModelに登録
+        model.addAttribute("status", HttpStatus.INTERNAL_SERVER_ERROR);
+
+        return "error";
+    }
+}
