+++ portal/src/main/java/com/portal/z/user/controller/userController.java
@@ -14,6 +14,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
@@ -33,7 +34,7 @@ import com.portal.z.common.domain.service.UserroleService;
 import com.portal.z.common.domain.service.RoleService;
 import com.portal.z.login.domain.model.AppUserDetails;
 
-
+@Transactional
 @Controller
 public class userController {
 
@@ -193,14 +194,19 @@ public class userController {
         
         // 環境マスタに登録したロール名（一般ユーザ）のrole_idを取得
         Role role = roleService.selectRoleid("ROLE_NAME_G");
-        
+        //
+        //ToDo
+        //ここでrole取得結果を評価しないといけない
+        //環境マスタに未登録の時やrole_idが取れなかったときは以降の処理を中止するなど
+        //
+
         userrole.setUser_id(form.getUser_id());               //ユーザーID
         userrole.setRole_id(role.getRole_id());               //ロールID
         
         // ユーザー登録処理
         boolean result_1 = userService.insert(user);
         boolean result_2 = userroleService.insert(userrole);
-
+        		
         // ユーザー登録結果の判定
         if (result_1 == true && result_2 == true ) {
         	model.addAttribute("result", "登録成功");
@@ -304,24 +310,15 @@ public class userController {
         
         user.setUpdate_user(user_auth.getUsername());         //更新者
 
-        try {
+        //更新実行
+        boolean result = userService.updateOne(user);
 
-            //更新実行
-            boolean result = userService.updateOne(user);
-
-            if (result == true) {
-                model.addAttribute("result", "更新成功");
-                System.out.println("更新成功");
-            } else {
-                model.addAttribute("result", "更新失敗");
-                System.out.println("更新失敗");
-            }
-
-        } catch(DataAccessException e) {
-
+        if (result == true) {
+            model.addAttribute("result", "更新成功");
+            System.out.println("更新成功");
+        } else {
             model.addAttribute("result", "更新失敗");
             System.out.println("更新失敗");
-
         }
 
         //ユーザー一覧画面を表示
@@ -365,7 +362,6 @@ public class userController {
         //ユーザー一覧画面を表示
         return getUserList(model);
     }
-
 
     /**
      * DataAccessException発生時の処理メソッド.
